refactor(effects): organize screen-space passes
